+++ evaluacion21-05-2025/ejercicio1/evaluador.py
@@ -1,0 +1,20 @@
+from utilidades import es_operador
+def evaluar_postfija(exp):
+    pila = []
+    for c in exp:
+        if c.isdigit():
+            pila.append(int(c))
+        elif es_operador(c):
+            b = pila.pop()
+            a = pila.pop()
+            if c == '+':
+                pila.append(a + b)
+            elif c == '-':
+                pila.append(a - b)
+            elif c == '*':
+                pila.append(a * b)
+            elif c == '/':
+                pila.append(a // b)
+            elif c == '^':
+                pila.append(a ** b)
+    return pila[0]
